i386: fix the major problems at least

+++ Indexing/LiteralSubstitutionTree.cpp
@@ -55,7 +55,7 @@ SLQueryResultIterator LiteralSubstitutionTree::getVariants(Literal* query, bool 
 SLQueryResultIterator LiteralSubstitutionTree::getAll()
 {
   return pvi(
-        iterTraits(getRangeIterator((unsigned long)0, _trees.size()))
+        iterTraits(getRangeIterator((size_t)0, _trees.size()))
          .flatMap([this](auto i) { return LeafIterator(&_trees[i]); })
          .flatMap([](Leaf* l) { return l->allChildren(); })
          .map([](const LeafData& ld) { return SLQueryResult(ld.literal, ld.clause); })
